Change facets to trials
Realized that Spacey is *not* a flashcard system.  It is
a review prioritization system.  It doesn't care about
how you are testing.  It just cares wether you got it correct
or not.

So now a fact has several trials.  The idea is that you
try one of the trials for each scheduled fact in each
session.  The trial that is chosen is dependent upon it's
priority.

I'm having difficulty deciding whether or not to add any
UI to this code.  I think I will not and instead write an
example flashcard system.