Move and rename things
I moved things around so that the source code is in src,
the test code is in spec, and the fakes/whatever are in
spec_helper.

I also renamed the Spacey class to be Facts because Spacey isn't
a thing; it's a namespace.  The thing that used to be called
Spacey is just a collection of Facts at the moment.  It will
probably be renamed at some point.